Mapas

Description

Tipos de dados do multiplayer e mapas para a realização do projeto de LI1 em 2021/22.

Tipo do multiplayer

data Jogo'

Tipo do multiplayer

Constructors

Jogo'

Fields

Mapa

Mapa normal

Jogador

Jogador1

Jogador

Jogador2

Funções para listas de jogos

allGameMaps :: Int -> [Jogo]

n mapas singleplayer, do mais facil para o mais dificil

allMultiMaps :: Int -> [Jogo']

n mapas multiplayer, do mais facil para o mais dificil

Número de mapas atuais e função que os cria

nMapaDouble :: Int

Número de mapas multiplayer

nMapaNormal :: Int

Número de mapas singleplayer

multiMap :: Int -> Jogo'

Recebe o nível e dá um jogo multiplayer

gameMap :: Int -> Jogo

Recebe o nível e dá um jogo