Tarefa1_2021li1g040

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2021/22.

Documentation

validaPotencialMapa

Arguments

:: [(Peca, Coordenadas)]
Recebe um mapa

-> Bool
Diz se o pudemos usar

Validação do Mapa

Para a validarção de um possivel mapa temos em conta cinco parametros sendo que, se algum deles tiver valor negativo, esse mapa será vetado. (a validação não necessita a organização dos termos).

A função validaPotencialMapa é definida da seguinte forma:

validaPotencialMapa :: [(Peca, Coordenadas)] -> Bool
validaPotencialMapa [] = False
validaPotencialMapa l = validaPosicao l && validaChao l && validaPorta l && validaPorta' r && validaCaixa l && validaEspacosVazios r
    where r = pura l (cMaisBaixos (percursoChao l))

E, com ela, validamos então:

A posição de cada peça (não haver sobreposição de peças) com a função validaPosicao;

A existencia de uma única porta (e esta ser a cima do chão)com as funções validaPorta e validaPorta';

A ausência de caixas a flutuar com a função validaCaixa;

Se há algum espaço vazio no mapa com a função validaEspacosVazios;

Se o mapa tem um chão completo com a função validaChao.

percursoChao

Arguments

:: [(Peca, Coordenadas)]
Recebe o mapa

-> [Coordenadas]
devolve as coordenadas por onde passa o chão

Coordenadas por onde passa o chão

listadoChao

Arguments

:: Int
Recebe uma direção

-> Coordenadas
Recebe as coordenadas do chão atual

-> [Coordenadas]
Recebe uma lista vazia que serve de acomulador

-> [Coordenadas]
Recebe a lista de coordenadas dos blocos do mapa

-> Coordenadas
Recebe as coordenadas do ultimo bloco

-> [Coordenadas]
Devolve a lista de blocos por onde passa o chão

Semelhante á função verifCont pois utiliza as funções habloco, eomaisbaixo, listaBlocos, sendo a única exceção a devolução de um mapa ao invês de um booliano.

validaPosicao

Arguments

:: [(Peca, Coordenadas)]
Recebe o mapa

-> Bool
Devolve False se dois termos estiverem na mesma posição

Validação das Posições

Para validar a posição comparamos, com uma função auxiliar, o primeiro termo com todos os outros (com a função recursiva validaCabeca) e, caso algum seja igual, devolve False.

Caso o termo não se repita, repete-se o procedimento para o termo asseguir.

validaPosicao :: [(Peca, Coordenadas)] -> Bool
validaPosicao [(_,_)] = True 
validaPosicao (termo : resto) = validaCabeca (termo : resto) && validaPosicao resto

validaCabeca [(p, (coluna1 , linha1))] = True 
validaCabeca ((p, (coluna1 , linha1)) : (_, (coluna2 , linha2)) : resto') = (coluna1 /= coluna2 || linha1 /= linha2) && validaCabeca((p, (coluna1 , linha1)) : resto')

Com isto verificamos se existem elementos em posições repetidas.

validaPorta

Arguments

:: [(Peca, Coordenadas)]
Recebe o mapa

-> Bool
Devolve True se existir uma única porta

Validação da Porta

Para validar as portas, começamos por correr a função até ocorrer a primeira porta, se tal não acontecer devolve False.

Caso aconteça é ativado uma segunda função chamada validaPorta' que continua a correr a função até encontrar uma segunda porta, se encontrar devolve False e se acabar o mapa devolve True.

Depois ainda é corrida uma função para verificar se a porta que existe está acessível

validaPorta :: [(Peca, Coordenadas)] -> Bool
validaPorta [] = False
validaPorta ((p,_) : resto) | p /= Porta = validaPorta resto
                            | otherwise = validaPorta'' resto
    where
    validaPorta'' :: [(Peca, Coordenadas)] -> Bool
    validaPorta'' [] = True
    validaPorta'' ((p,_) : resto) =  p /= Porta && validaPorta'' resto

validaPorta'

Arguments

:: [(Peca, Coordenadas)]
Recebe o mapa com o chão coberto até à última linha

-> Bool
devolve True se ainda existir porta

validaPorta' :: [(Peca, Coordenadas)] -> Bool
validaPorta' [] = False
validaPorta'  ((p,_) : resto) | p /= Porta = validaPorta' resto
                              | otherwise = validaPorta'' resto
    where
    validaPorta'' :: [(Peca, Coordenadas)] -> Bool
    validaPorta'' [] = True
    validaPorta'' ((p,_) : resto) =  p /= Porta && validaPorta'' resto

Com isto validamos a porta.

validaCaixa

Arguments

:: [(Peca, Coordenadas)]
Recebe um mapa

-> Bool
Devolve True se não existirem caixas ou se, as que existirem tiverem um Bloco/Caixa por baixo

Validação das caixas

Não é necissária a existência de Caixas num mapa, porém, é necessário um chão por baixo delas.

Tendo isso em conta, procede-se então com a função validaCaixa' que precorre o mapa em busca de uma caixa, se esta não existir não há problemas (devolve True).

Caso exista, com a função validaSubCaixa, vai-se procurar no mapa dado, a base da caixa, se esta existir e não for uma porta devolverá True.

validaCaixa :: [(Peca, Coordenadas)] -> Bool
validaCaixa l = validaCaixa' l l
    where
    validaCaixa' :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)] -> Bool
    validaCaixa' [] _= True
    validaCaixa' ((peca,(coluna,linha)) : ss) mapa | peca == Caixa = validaSubCaixa (coluna,linha + 1) mapa && validaCaixa' ss mapa
                                                   | otherwise = validaCaixa' ss mapa
        where
        validaSubCaixa :: Coordenadas -> [(Peca, Coordenadas)] -> Bool
        validaSubCaixa (_,_) [] = False
        validaSubCaixa (col1, line1) ((peca,(col2, line2)): ss) 
                        | col1 == col2 && line1 == line2 && (peca == Caixa || peca == Bloco) = True
                        | otherwise = validaSubCaixa (col1, line1) ss

Com isto validamos as possiveis caixas.

validaEspacosVazios

Arguments

:: [(Peca, Coordenadas)]
Recebe o mapa com a baixo do chão coberto de blocos

-> Bool
Devolve True se o mapa tiver pelo menos um espaço vazio e False se tal não acontecer

Verificação da existência de espaços vazios

Para o mapa ser usável é necessário que haja pelo menos um sitio onde colocar a personagem.

Para isso utilizamos a função validaEspacosVazios que compara o espaço ocupado por peças com o tamanho escalar do mapa caso tal espaço seja igual ou superior devolve False.

validaEspacosVazios l = validaEspacosVazios' l (tamanhoMapa l)
    where
    validaEspacosVazios' :: [(Peca, Coordenadas)] -> (Int,Int) -> Bool
    validaEspacosVazios' p (c,l) = elemVazios p || espacoOcupado < espacosTotais
        where espacoOcupado = length p
              espacosTotais = (c+1)*(l+1)
              elemVazios [] = False
              elemVazios ((pecas, (_, _)) : xs) | pecas == Vazio = True
                                                | otherwise = elemVazios xs

Com isto verificamos se o mapa é jogável.

validaChao

Arguments

:: [(Peca, Coordenadas)]
Recebe o mapa

-> Bool
Verifica se existe caixas e se existir verifica se não flotuam

Verificação da existência de um chão

Para a personagem poder percorrer todo o mapa é necessário que exista um chão para ela andar. Verificamos tal existência com o auxilio da da função validaChao que percorre a função e verifica se existe umchão continuo que passa por todas as colunas, se tal existir devolve true.

Estes testes são exemplos:

testes :: Int -> Bool
testes x |x== 1 = validaChao [(Bloco,(0,3)),(Porta,(0,2)),(Bloco, (1,3)),(Bloco, (1,4)),(Bloco,(1,5)),(Bloco,(2,5)),(Bloco,(3,5)),(Bloco,(3,4)),(Bloco,(3,3)),(Bloco,(4,3)),(Bloco,(5,3))] --True
         |x== 2 = validaChao [(Bloco,(0,3)),(Porta,(0,2)),(Bloco, (0,4)),(Bloco, (1,3)),(Bloco,(1,4)),(Bloco,(1,5)),(Bloco,(2,5)),(Bloco,(3,5)),(Bloco,(3,4)),(Bloco,(3,3)),(Bloco,(4,3)),(Bloco,(5,3))] --True
         |x== 3 = validaChao [(Bloco,(0,3)),(Porta,(0,2)),(Bloco, (1,3)),(Bloco, (1,4)),(Bloco,(1,5)),(Bloco,(2,5)),(Bloco,(3,5)),(Bloco,(3,4)),(Bloco,(3,3)),(Bloco,(4,3)),(Bloco,(5,3)),(Bloco, (0,0)),(Bloco, (3,0))]--True
         |x== 4 = validaChao [(Bloco,(0,3)),(Porta,(0,2)),(Bloco, (1,3)),(Bloco, (1,4)),(Bloco,(1,5)),(Bloco,(2,5)),(Bloco,(3,5)),(Bloco,(3,4)),(Bloco,(3,3)),(Bloco,(4,3)),(Bloco,(5,3)),(Bloco, (0,0)),(Bloco, (3,0)),(Bloco,(2,6))] --False
         |x== 5 = validaChao [(Bloco,(0,0)),(Bloco,(2,0))] --False
         |x== 6 = validaChao [(Bloco,(1,1))]--False 
         |x== 7 = validaChao [(Bloco,(0,4))]--True
         |x== 8 = validaChao [(Bloco,(0,4)),(Bloco,(0,3)),(Bloco,(0,2)),(Bloco,(1,2))] --True
         |x== 9 = validaChao [(Bloco,(0,0)),(Bloco,(1,0)),(Bloco,(1,1)),(Bloco,(1,2)),(Bloco,(1,3)),(Bloco, (2,1))]--False
         |x== 0 = validaChao [(Bloco,(0,0)),(Bloco,(1,0)),(Bloco,(1,1)),(Bloco,(1,3)),(Bloco,(2,3))]--False
         |x==10 = validaChao [(Bloco,(0,3)),(Bloco, (1,3)),(Bloco, (2,3)),(Bloco, (1,4)),(Bloco,(1,5)),(Bloco,(2,5)),(Bloco,(3,5)),(Bloco,(4,4)),(Bloco,(4,3)),(Bloco,(4,5)),(Bloco,(5,3)),(Bloco, (0,0)),(Bloco, (3,0))] --True
         |x==11 = validaChao [(Bloco,(0,0)),(Bloco,(1,1)),(Bloco,(2,2)),(Bloco,(3,1)),(Bloco,(4,0))] --True

Com isto confirmamos por fim se existe algum boraco no mapa.

Para facilitar o processo começamos,com listaBlocos, por isolar as coordenadas dos Blocos.

listaBlocos

Arguments

:: [(Peca, Coordenadas)]
Recebe o mapa completo

-> [Coordenadas]
Devolve uma lista com apenas as coordenadas de todos os blocos

coordenadas de todos os blocos

verifCont

Arguments

:: Int
De onde veio

-> Coordenadas
Onde está de momento

-> [Coordenadas]
coordenadas dos blocos do mapa

-> Coordenadas
Tamanho do mapa

-> Bool

A função verifCont recebe um inteiro que pode ser um de três e representa a direção por onde veio:

0 -> vem da esquerda

1 -> vem de cima

2 -> vem de baixo

Precisamos de saber em que coordenadas começar o chão e onde acabar, para isso usamos as funções que se seguem.

findFrstChao

Arguments

:: [Coordenadas]
Recebe a lista com as coordenadas dos blocos

-> Coordenadas
Recebe as coordenadas do primeiro bloco do ma

-> Coordenadas
Devolve as coordenadas do bloco mais perdo do canto inferior esquerdo

Função que verifica se a coordenada recebida é a mais baixa da coluna.

findLastChao

Arguments

:: [Coordenadas]
Recebe a lista com as coordenadas dos blocos

-> Coordenadas
Recebe as coordenadas do primeiro bloco do mapa

-> Coordenadas
Devolve as coordenadas do bloco mais perdo do canto inferior direito

procura o ultimo chão

eomaisbaixo

Arguments

:: Coordenadas
Recebe as coordenadas a verificar

-> [Coordenadas]
Recebe a lista com as coordenadas dos blocos

-> Bool
Verifica se a coordenada recebida é a mais baixa da coluna

verifica se é o mais baixo da coluna

habloco

Arguments

:: Coordenadas
Recebe as coordenadas a verificar

-> [Coordenadas]
Recebe a lista com as coordenadas dos blocos

-> Bool
Verifica se existe um bloco na coordenada que recebe

Função que verifica se existe um bloco na coordenada que recebe.

tamanhoMapa

Arguments

:: [(Peca, Coordenadas)]
Recebe o mapa

-> (Int, Int)
Devolve o par (ultimaColuna, ultimaLinha)

Para o quarto e quinto passo foi usado uma função que devolvia a escala do mapa. A essa função chamamos tamanhoMapa.

tamanhoMapa :: [(Peca, Coordenadas)] -> (Int,Int)
tamanhoMapa  = maximo (0,0)
    where
    maximo :: (Ord a, Ord b) => (a, b) -> [(a3, (a, b))] -> (a, b)
    maximo acc [] = acc
    maximo (mc , ml) ((p1, (c1 , l1)) : resto) | mc >= c1 && ml >= l1 = maximo (mc,ml) resto
                                               | mc >= c1 && ml < l1 = maximo (mc,l1) resto
                                               | mc < c1 && ml >= l1 = maximo (c1,ml) resto
                                               | mc < c1 && ml < l1 = maximo (c1,l1) resto

Através desta função é nos fornecido tanto o número máximo de linhas/colunas como o número de parcelas do mapa.