Tarefa2_2021li1g040

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2021/22.

Construção e desconstrução do mapa

A função constroiMapa recebe numa lista de pecas, ordenada ou não, e retorna um mapa, ou seja, uma lista de listas de pecas, em que cada lista de pecas corresponde a uma linha do mapa. A função desconstroiMapa recebe um mapa e torna-a numa lista de peças.

Documentation

constroiMapa

Arguments

:: [(Peca, Coordenadas)]
Recebe a lista de pecas, ordenadas ou não

-> Mapa
Retorna o mapa construido

constroiMapa :: [(Peca, Coordenadas)] -> Mapa                 
constroiMapa input_mapa | validaPotencialMapa input_mapa = constroiMapaT4 input_mapa
                        | otherwise = []

A função constroiMapa recorre uma função auxiliar controiMapaT4 pois, caso recorressemos à função constroiMapa na tarefa 4, esta poderia falhar devido à validação do mapa. Sendo assim, a verificação do mapa é feita apenas em constroiMapa procedendo para a construção do mapa em constroiMapaT4 (caso o validaMapa retorne True).

constroiMapaT4 :: [(Peca, Coordenadas)] -> Mapa

desconstroi mapa da T4

ordena_linha_coluna

Arguments

:: [(Peca, Coordenadas)]
Recebe um acumulador

-> [(Peca, Coordenadas)]
Recebe a lista de pecas que serao verificas

-> [(Peca, Coordenadas)]
Recebe a mesma lista de peças que nunca será alterado, servindo apenas para repor a lista de peças anterior quando a função passa para uma nova linha

-> Int
Recebe o número de linhas do mapa

-> [(Peca, Coordenadas)]
Retorna a lista de peças ordenada

A função ordena_linha_coluna ordena o mapa por linhas e consecutivamente por colunas, recorrendo posteriormente à função ordenaColuna para o fazer retornando assim uma lista de peças ordenada tornando mais facil a execução da função constroiMapa

ordenaColuna

Arguments

:: [(Peca, Coordenadas)]
Recebe a lista de peças ordenada por linhas

-> [(Peca, Coordenadas)]
Retorna a lista de peças ordenada por linhas e por colunas

A função ordenaColuna funciona similarmente ao quicksort mas ligeiramente adaptado ao problema em questão recorrendo às funções maisPequenos e maiores

maisPequenos :: Int -> [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

Coleciona as pecas com a coluna mais pequena em comparação com a do elemento central da função ordenaColuna

maiores :: Int -> [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

Coleciona as pecas com a coluna maior em comparação com a do elemento central da função ordenaColuna

isCoordIn :: Coordenadas -> [(Peca, Coordenadas)] -> Bool

A isCoordin verifica se a coordenada em questão consta no input

nLinhas :: [(Peca, Coordenadas)] -> Int

Calcula o número de linhas do mapa

desconstroiMapa

Arguments

:: Mapa
Recebe o mapa

-> [(Peca, Coordenadas)]
Retorna uma lista de peças com as respetivas coordenadas

A função desconstroiMapa recebe uma lista de lista de peças e retorna uma lista de peças com as suas respetivas coordenadas ignorando os espaços vazios.

desconstroiMapa' :: Mapa -> (Int, Int) -> [(Peca, Coordenadas)]

auxiliar da desconstroi mapa