Tarefa4_2021li1g040

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2021/22.

Movimentos do Jogador

A ação de movimentação de um jogador, com a função moveJogador, baseia-se na restrição de movimentos possiveis para cada situação em que o jogador se encontra e o respetivo movimento que recebe. Quando o o movimento aplicado é:

Movimento AndarEsquerda e AndarDireita: a função verifica todos os casos em que o jogador se pode encontrar ao correr este movimento. Casos em como o jogador tem um bloco a bloquear o seu deslocamento (quando tem ou não caixa), ou quando o jogador se desloca de uma posicao mais alta para uma baixa obrigando-o a ter que descer para a coordenada mais abaixo possível.

Movimento Trepar : Tal como nos movimentos anteriores, a função verifica os casos em o jogador o pode fazer (quando tem ou não tem caixa).

Movimento InterageCaixa : Neste caso a função moveJogador vai analisar se o jogador já transporta caixa. Caso não o faça, vai verificar, primeiro de tudo, se existe caixa à frente do jogador e se o ambiente à sua volta permite que o jogador pegue nessa mesma caixa. Caso o jogador já transporte uma caixa a função vai verificar se pode pousar a caixa e vai pousa-la na coordenada mais abaixo possível (por outras palavras, vai deixar a caixa "cair")

Estando feita a função moveJogador, para fazer a correrMovimentos foi simplesmente preciso utilizar a recursividade com a função moveJogador.

Documentation

correrMovimentos

Arguments

:: Jogo
Recebe o jogo inicial

-> [Movimento]
Recebe a lista de movimentos

-> Jogo
Retorna o jogo final aplicando a lista de movimentos ao jogo

Função princilal do corre movimentos

moveJogador

Arguments

:: Jogo
Recebe o jogo

-> Movimento
Recebe o movimento a aplicar

-> Jogo
Retorna o jogo com o movimento aplicado

A função 'moveJogador recorre a uma função auxiliar moveJogador' que aplica a função posicionaBonecoCaixa que vai posicionar o boneco na coordenada mais a baixo possivel de onde se encontra

posicionaFinal :: Jogo -> Jogo

Posição final

moveJogador' :: Jogo -> Movimento -> Jogo

Funções auxiliares

verifExisteCaixaOeste

Arguments

:: Mapa
Recebe o mapa

-> Coordenadas
Recebe as coordenadas do jogador

-> Bool
Retorna True se existir caixa

Verifica se existe caixa no bloco em frente do boneco no lado Oeste

verifExisteCaixaEste

Arguments

:: Mapa

-> Coordenadas

-> Bool
Retorna True se existir caixa

Verifica se existe caixa no bloco em frente do boneco no lado Este

verifTreparEsteComCaixa

Arguments

:: Mapa
Recebe o mapa

-> Coordenadas
Recebe as coordenadas do jogador

-> Bool
Retorna True se poder trepar

Verifca se o boneco pode trepar com caixa no sentido Este

verifTreparEsteSemCaixa :: Mapa -> Coordenadas -> Bool

Verifica se o boneco pode trepar sem caixa nos sentido Este

verifTreparOesteComCaixa :: Mapa -> Coordenadas -> Bool

Verifca se o boneco pode trepar com caixa no sentido Oeste

verifTreparOesteSemCaixa :: Mapa -> Coordenadas -> Bool

Verifica se o boneco pode trepar sem caixa nos sentido Oeste

verifCaixaOeste

Arguments

:: Mapa
Recebe o mapa

-> Coordenadas
Recebe as coordenadas do jogador

-> Bool
Se tiver algo a obstruir retorna True

Verifica se o boenco pode andar com caixa para Oeste

verifCaixaEste :: Mapa -> Coordenadas -> Bool

Verifica se o boneco pode andar coma caixa para Este

verifEste

Arguments

:: [(Peca, Coordenadas)]
Recebe o mapa

-> Coordenadas
Recebe as coordenadas do jogador

-> Bool
Se nao tiver algo a obstruir retorna True

Verifica se o boneco pode andar no sentido Este

verifOeste

Arguments

:: [(Peca, Coordenadas)]
Recebe o mapa

-> Coordenadas
Recebe as coordenadas do jogador

-> Bool
Se nao tiver algo a obstruir retorna True

Verifica se o boneco pode andar no sentido Oeste

verifJogador

Arguments

:: Mapa
Recebe o mapa

-> Coordenadas
Recebe as coordenadas do jogador

-> Bool
Retorna True se o boneco não poder ser inserido

Se retornar True o boneco nao pode ser inserido no mapa (se tudo funcionar bem so podera retornar True no primeio movimento)

posicionaBonecoCaixa

Arguments

:: Mapa
Recebe o mapa

-> Coordenadas
Recebe as coordenadas inicias do jogador/caixa

-> Coordenadas
Retorna as coordenadas do jogador/caixa posicionados no sitio certo

Funcao para quando o boneco andar, va sempre para o bloco mais abaixo de onde andou (i.e. quando cai), ou quando é largada uma caixa, tendo o mesmo efeito

posicionaCaixa

Arguments

:: Direcao
Recebe a direcao do jogador

-> Coordenadas
Recebe as coordenadas do jogador

-> Mapa
Recebe o mapa

-> Mapa
Retorna o mapa alterado

Esta funcao recebe a direcao do jogador, as suas coordenadas e o mapa e posicona a caixa na linha mais em baixo possivel para a coluna em questão

apanhaCaixa

Arguments

:: Direcao
Recebe a direcao do jogador

-> Coordenadas
Recebe as coordenadas do jogador

-> Mapa
Recebe o mapa do jogo

-> Mapa
Retorna o mapa alterado

Se o jogador pegar numa caixa, verifica se a pode pegar e retira-a do mapa, caso contrario o mapa fica igual

mapaLargaCaixa

Arguments

:: Coordenadas
Recebe as coordenadas do jogador

-> Mapa
Recebe o mapa do Jogo

-> Direcao
Recebe a direcao

-> Mapa
Retorna o mapa resultado de mapaLargaCaixa'

Quando o jogador está a segurar uma caixa e ela nao esta realmente no mapa, esta funcao volta a po-la no mapa quando o caixa==True e o movimento é InterageCaixa

mapaLargaCaixa'

Arguments

:: Coordenadas
Recebe as coordenadas do jogador

-> Mapa
Recebe o mapa

-> Direcao
Recebe a direcao do jogador

-> Mapa
Retorna o mapa alterado

Esta funcao faz o mesmo de a de cima so que quando o boneco tenho uma caixabloco em frente ent poe em cima dessa caixabloco

nColunas

Arguments

:: [(Peca, Coordenadas)]
Recebe o mapa do jogo

-> Int
Retorna o numero de colunas do mapa

numero de colunas do mapa