Tarefa6_2021li1g040

Description

Módulo para a realização da Tarefa 6 do projeto de LI1 em 2021/22.

BOT

O objetivo final desta tarefa seria encontrar uma forma de automatizar a resolução de um mapa, recebendo apenas um número máximo de movimentos. Inicialmente foi tentada uma abordagem de encontrar uma forma geral da resolução dos mapas, por exemplo, andar como jogador na direção da porta e caso não seja possível, voltar atrás e encontrar caixas e voltar a tentar resolver. Obviamente, que dada a complexidade e diversidade que os mapas podem ter, esta não seria uma forma viável de fazer este bot, como muitas outras formas pensadas ao estilo desta. Haveriam certamente várias maneiras de o resolver mas, a forma utilizada, neste caso, explicada de uma forma sucinta, foi através do calculo dos caminhos possíveis dentro do limite de movimentos fornecidos pelo utilizador. Obviamente que esta forma apresenta um ponto fraco que é a eficiência do programa, tendo sido este o maior desâfio da tarefa, ou seja, tentar arranjar maneiras de reverter esta ineficiência.

Para resolver este bot, comecei por fazer um construtor de uma árvore com quatro ramos (BTree), um para cada movimento diponível, que receberia inicialmente os Parametros iniciais do jogo e, partir daí, construir uma árvore, com várias funções auxiliares, dentro dos limites fornecidos. Para encurtar a árvore o máximo possível, foram inseridas uma série de medidas para terminar com ramos que tivessem movimentos inúteis, como por exemplo, andar contra um bloco. A par disto, foi criado um sistema de flags, em que um movimento inútil causaria a uma variável nos Parametros ser alterada para 1, e se dois movimentos seguidos fossem inuteis, o ramo seria cortado ali.

Inicialmente a função resTConstroi retornava uma árvore completa, mas para facilitar a função limiter, a função passou a retornar apenas as folhas da árvore dentro de uma lista. A função limiter, recebe o número de movimentos totais, e em vez de correr a resTConstroi de uma vez só, executa-a movimento a movimento para permitir que ela façaa verificação de que se há ou não um ramo da árvore que já tenha encontrado a solução do mapa. Desta forma, ao receber 100 movimentos para um jogo que seria facilmente resolvido em 10, o bot iria parar de encontrar uma solução depois desse ponto.

Uma outra observação feita foi que, se um mapa não tiver caixas, ou as que tiver forem inutilizáveis, não valeria a pena fazer com que o bot calculasse resoluções possíveis onde integrasse o movimento InterageCaixa, resultando na função verifCaixaValida.

Documentation

data RTree a

A RTree corresponde ao formato da árvore de resolução onde cada nodo filho vai corresponder a um movimento (AndarEsquerda,InterageCaixa,Trepar,AndarDireita)

Constructors

Empty

Node a (RTree a) (RTree a) (RTree a) (RTree a)

type ResTree = RTree Parametros

A RTree é uma árvore em que recebe os diferentes Parametros

type Parametros = (Movimento, NMoves, Jogo, InPorta, LMoves, Flag, Caixas)

Os Parametros são informações, mutáveis ou não, que cada nodo deve levar consigo

type Flag = Int

A Flag é do tipo Int e é utilizada para avisar o codigo da existência de movimentos inúteis

type NMoves = Int

O NMoves é um contador do tipo Int

type InPorta = Bool

O InPorta é um bool para avisar o código se o jogador já chegou à porta ou não

type LMoves = [Movimento]

O LMoves corresponde a uma lista de movimentos

type Caixas = Bool

O caixas é do tipo Bool e não é mutável durante o resTConstroi, servindo para alertar se o jogo tem caixas utilizáveis ou não

resolveJogo :: Int -> Jogo -> Maybe LMoves

A função resolveJogo é a função principal desta tarefa, e vai receber o número de movimentos máximos, e o Jogo. A partir daí, vai ativar a função limiter

resTConstroi :: Int -> ResTree -> [ResTree]

A função resTConstroi é a função responsável por calcular a sequência de movimentos, deixando de fora os ramos que contenham movimentos inúteis a uma possível resolução.

isInPorta :: Coordenadas -> Mapa -> Int -> Bool

A função isInPorta faz a verificação de se o jogador já chegou a porta, retornando True caso isto se verifique. Ele recebe as coordenadas do jogador, o mapa e o terceiro parametro que recebe é um acumulador.

isInPorta' :: Int -> [Peca] -> Int -> Bool

Função auxiliar de isInPorta

verifCaixaValida :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)] -> Bool

Esta funcao retorna True se existirem caixas no mapa e se essas caixas forerm utilizaveis. A função recebe o mapa duas vezes, visto que um deles serverá apenas para passar às funções auxiliares verifCaixaCima e verifCaixaLados

verifCaixaCima :: [(Peca, Coordenadas)] -> Coordenadas -> Bool

Verifica se existem blocos por cima das caixas

verifCaixaLados :: [(Peca, Coordenadas)] -> Coordenadas -> Int -> Bool

Verifica se existem dois blocos ao lado da caixa, um em cada lado

limiter :: Int -> [ResTree] -> Maybe LMoves

A função limiter é chamada em resolveJogo e recebe o limite de movimentos imposto nela e o primeiro nodo com o dados iniciais do jogo. Esta função em vez de inserir o limite de movimentos todos de uma vez na função resolveJogo, permite apenas que a função calcule um movimento, de forma a poder analisar a lista fornecida e ver se encontrou a resolução do jogo ou não.

checker :: [ResTree] -> Bool

Esta função verifica se a resolução do jogo está presente na lista de nodos

checker' :: ResTree -> Bool

Função auxiliar de checker

movesExtractor :: [ResTree] -> LMoves

Quando uma resolução é encontrada, esta função é chamada, para poder retirar a lista de movimentos do nodo que retornou True para o inPorta

recursiveResT :: [ResTree] -> [ResTree]

A recursiveResT aplica o resTConstroi à lista de nodos permitindo apenas um movimento máximo, retornando uma lista de nodos.